+++ client/lib/ft4/admin/admin-operations.ts
@@ -6,6 +6,7 @@ import {
 } from "/ft4/accounts/auth-descriptor/types";
 import { BufferId } from "/cryptoUtils";
 import { Amount } from "../asset/interfaces";
+import { Asset } from "../asset/types";
 
 export function registerAccount(
   authDescriptor: GtvAuthDescriptorRegistration<GtvAuthDescriptorArgs>,
@@ -43,5 +44,20 @@ export function mint(
     formatter.ensureBuffer(accountId),
     formatter.ensureBuffer(assetId),
     amount.value,
+  );
+}
+
+export function registerCrosschainAsset(
+  asset: Asset,
+  originBrid: BufferId,
+): Operation {
+  return op(
+    "ft4.admin.register_crosschain_asset",
+    asset.name,
+    asset.symbol,
+    asset.decimals,
+    asset.brid,
+    asset.iconUrl,
+    formatter.ensureBuffer(originBrid),
   );
 }