+++ client/lib/ft4/authentication/ft/key-handler.ts
@@ -52,7 +52,7 @@ async function sign(
   transaction.signatures.push(
     await keyStore.sign(
       gtx.getDigestToSign({
-        blockchainRID: transaction.blockchainRID,
+        blockchainRid: transaction.blockchainRid,
         signers: transaction.signers,
         operations: transaction.operations,
       }),