+++ client/lib/ft4/crosschain/errors.ts
@@ -1,0 +1,8 @@
+export class OrchestratorError extends Error {
+  type: string;
+
+  constructor(message: string, type = "OrchestratorError") {
+    super(message);
+    this.type = type;
+  }
+}