+++ client/lib/ft4/crosschain/query-functions.ts
@@ -1,0 +1,11 @@
+import { BufferId } from "../../cryptoUtils";
+import { assetOriginById } from "./queries";
+import { Connection } from "../types";
+import { Buffer } from "buffer";
+
+export async function getAssetOriginById(
+  connection: Connection,
+  id: BufferId,
+): Promise<Buffer | null> {
+  return await connection.query<Buffer>(assetOriginById(id));
+}