+++ client/lib/ft4/events/index.ts
@@ -1,3 +1,4 @@
+export { EventHandlers } from "./handlers";
 export type { Listener } from "./types";
 export { ftEventEmitter } from "./ftEventEmitter";
 export { EventEmitter } from "./emitter";