+++ client/lib/ft4/utils/types.ts
@@ -45,7 +45,7 @@ export type PaginatedEntity<T> = {
 };
 
 export type TxBuilderTransaction = {
-  blockchainRID: Buffer;
+  blockchainRid: Buffer;
   operations: RellOperation[];
   signers: Buffer[];
   signatures: Buffer[];
@@ -63,3 +63,7 @@ export interface RellOperationStructure {
   mount: string;
   parameters: any[];
 }
+
+export type RequireTogether<T, Keys extends keyof T> = T & {
+  [K in Keys]-?: T[K];
+};